+++ database/release/006_seed_feriados_base.sql
@@ -5,7 +5,7 @@ Objetivo: cargar reglas base de feriados irrenunciables.
 No carga feriados por fecha ni consulta APIs externas.
 */
 
-USE APP_SCHEDULER_TEST_INSTALL;
+USE [$(DB_NAME)];
 GO
 
 MERGE dbo.reglas_feriados_irrenunciables AS destino